Allowed for incremental

+++ app/modules/github.py
@@ -15,7 +15,7 @@ def get_plans_from_github():
         logging.error(f'[get_plans_from_github] Could not pull plans from [{repo_url}]. {e}')
         return { "status" : "error", "info" : f"{e}" }
     
-def update_plans_from_github():
+def update_plans_from_github(force=False):
     if not os.listdir(plan_path): # Check if the plans currently exist
         logging.warn(f"[update_plans_from_github] No plans found. Downloading plans from [{repo_url}]")
         if "success" not in (result := get_plans_from_github()['status']): return result
@@ -26,7 +26,10 @@ def update_plans_from_github():
         #repo = Repo(plan_path)
         #repo.remotes.origin.pull()
         # TEMP SOLUTION [START] - as python git is full of sh**
-        execute_shell(cmd_as_string=f'cd {plan_path} && git reset --hard && git pull')
+        if force:
+            execute_shell(cmd_as_string=f'cd {plan_path} && git reset --hard && git pull')
+        else:
+            execute_shell(cmd_as_string=f'cd {plan_path} && git pull')
         # TEMP SOLUTION [END] - as python git is full of sh**
         return { "status" : "success" }
     except Exception as e:
